requirements.txt
streamlit==1.46.1
streamlit-lottie
requests
pillow